Add files via upload
Changes have touched the mechanism of positioning the window. Now you can move it across the entire page.
Implemented an automatic change of scale of the value indicator
Appended the ability to show not only values in the range from 0 to 100 percent, but also any number, is added in the header. For correct display, in this case you need to transfer the maximum value.